- update for token id

+++ rpcserver/rpcfunction.go
@@ -563,7 +563,7 @@ func (rpcServer RpcServer) handleEstimateFeeWithEstimator(params interface{}, cl
 	// param #2: numbloc
 	numblock := uint64(8)
 	if len(arrayParams) >= 3 {
-		numblock = arrayParams[2].(uint64)
+		numblock = uint64(arrayParams[2].(float64))
 	}
 
 	// param #3: tokenId
